libflux: create internal message library
Problem: We wish to move some messaging functions out of libflux, but there are some
shared structs / functions that will be needed if we do so.

Solution: Create a message convenience library within src/common/libflux
and add it to libflux-internal so it can be used internally by other flux-core code.

The following structs/functions were migrated out of message.c into their own
individual files.

struct flux_msg -> message_private.h
struct msg_iovec, msg_to_iovec(), iovec_to_msg() -> message_iovec.[ch]
PROTO macros, msg_proto_setup(), proto_get_u32() -> message_proto.[ch]
struct route_id, msg_route_push/append/clear/delete_last() -> message_route.[ch]

Update several linker dependencies to ensure all symbols are available for compilation.
