+++ src/common/libpmi/Makefile.am
@@ -50,12 +50,8 @@ test_ldadd = \
 	$(top_builddir)/src/common/libflux/libflux.la \
 	$(top_builddir)/src/common/libpmi/libpmi_client.la \
 	$(top_builddir)/src/common/libpmi/libpmi_server.la \
-	$(top_builddir)/src/common/libutil/libutil.la \
+	$(top_builddir)/src/common/libflux-internal.la \
 	$(top_builddir)/src/common/libtap/libtap.la \
-	$(top_builddir)/src/common/liblsd/liblsd.la \
-	$(top_builddir)/src/common/libtomlc99/libtomlc99.la \
-	$(top_builddir)/src/common/libev/libev.la \
-	$(top_builddir)/src/common/libczmqcontainers/libczmqcontainers.la \
 	$(ZMQ_LIBS) \
 	$(JANSSON_LIBS) \
 	$(LIBPTHREAD) \
